base: Introduce Extended Restart
Extended restart is a global actions framework, which ties
into the system global actions menu. This framework provides
various restart options. This is based up Advanced reboot from
nougat and below.

Instead of providing a basic AlertDialog like before, extended restart
utilizes the new power menu dialog in 8.1. We've created a minimal
forked class of the origin GlobalActionsDialog and added a hook into the
original restart button. Single pressing the restart button reboots the
system as normal, but long pressing it brings the extended fork into
view. This allows Extended Restart to be optional without the need
of creating toggles and system variables.

Extended Restart supports 3 restart options:

- Quick Reboot (Destroy AM, which will destory all services as a result,
                forcing a small restart)
- Recovery (Restart the device into the installed recovery mode)
- Bootloader (Restarts the device into the OEM's provided bootloader
              menu)

@Sirhc95: Ported from oreo-mr1 branch into Parfait. Extended Restart
no longer holds it's own class. Instead, create a version of handleShow()
to accomadate for the dialog. This limits the duplication of code in Global
Actions.

Change-Id: I2c6690e558d3cef2b0d7f7aa22265fc0cafb7bef