REAPPLY: SystemUI: Fix showing duplicated charging status
In commit 85e5a57 it was possible that plugging a device in,
then turning on the screen would show the charging status twice.
To fix, don't append the charging status to if there is a
transient notification.

(Originally 34e1765ecdc5896e7340cf2a3af825864d3dc085, accidentally dropped in r38 merge)

Change-Id: Ic031fbaf9a99d044f50dc65e2b189773222ebe80
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>